requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.100
ultralytics-thop==2.0.14
urllib3==2.3.0
